+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Version 2023-rc.2-dev
 -  fix pyxtp option restart (#1074)
 -  fix Changelog check (#1070)
 -  add python bindings for xtp (#1061)
+-  replaced pointers in MD2QM mapping (#1075)
 
 Version 2023-rc.2 (released 02.06.23)
 =====================================